requirements.txt
streamlit
googletrans==4.0.0rc1
pillow